--- data/photos.json
@@ -1,25 +1,0 @@
-[
-    { 
-        "id": "1",
-        "name": "Dropping the Heart Chute",
-        "description": "The heart chute is one of my favorite chutes",
-        "category": "ACTION",
-        "githubUser": "gPlake",
-        "created": "2018-04-15T19:09:57.308Z"
-    },
-    { 
-        "id": "2",
-        "name": "Enjoying the sunshine",
-        "category": "SELFIE",
-        "githubUser": "sSchmidt",
-        "created": "3-21-1987"
-    },
-    { 
-        "id": "3",
-        "name": "Gunbarrel 25",
-        "description": "25 laps on gunbarrel today",
-        "category": "LANDSCAPE",
-        "githubUser": "sSchmidt",
-        "created": "3-28-2018"
-    }
-]
