--- data/tags.json
@@ -1,6 +1,0 @@
-[
-    { "photoID": "1", "userID": "gPlake" },
-    { "photoID": "2", "userID": "sSchmidt" },
-    { "photoID": "2", "userID": "mHattrup" },
-    { "photoID": "2", "userID": "gPlake" }
-]
