--- data/users.json
@@ -1,5 +1,0 @@
-[
-    { "githubLogin": "mHattrup", "name": "Mike Hattrup" },
-    { "githubLogin": "gPlake", "name": "Glen Plake" },
-    { "githubLogin": "sSchmidt", "name": "Scot Schmidt" }
-]
