+++ typeDefs.graphql
@@ -42,8 +42,10 @@ type Query {
     me: User
     totalPhotos: Int!
     allPhotos: [Photo!]!
+    Photo(id: ID!): Photo
     totalUsers: Int!
     allUsers: [User!]!
+    User(login: ID!): User
 }
 
 type Mutation {
